include/drivers: Add RTC and RTCC API headers
The RTC API header support a real-time-counter, which is a 32-bit
binary counter with a frequency of 1 Hz. Each tick is correlated
with a second passing. It supports multiple individual alarms,
which compare with the RTC tick value.

The RTC and counter APIs have similarities, yet the counter API
can not cover the RTC since the RTC requires the ability to set
the tick value. The counter API also contains redundant functions
since the RTC top value can not be altered, and is known, and
the frequency is fixed, making the freq get function redundant.

The  real time clock calendar (RTCC) API support devices which
contain a gregorian or persian calendar, used to keep track of
date and time. These contain alarms which compare with fields
of the RTCC date and time, like weekday, hour and minute. They
return the "broken-down" date and time instead of an elapsed
time in seconds like RTCs do.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>