tests/drivers: Add RTC and RTCC API test suites
Tests have been added which test the behavior of RTC and
RTCC devices. The tests have been validated against the
emulated RTC and RTCC device drivers, implemented on the
native_posix board. The API tests will run on any board
which supports 'rtc' or 'rtcc' respectively in the boards
yaml file, and includes the alias 'rtc' ir 'rtcc'
respectively in the boards devicetree.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>