boards/posix/native_posix: Add emulated RTC and RTCC devices
An RTC and and RTCC emulated device using the emulated RTC and
RTCC device drivers has been added to the native_posix board
providing support for RTC and RTCC for the native_posix board.

These emulated drivers do not replace the native_rtc module.

native_rtc is used to control the simulated time, and to get the
time from the host system, contrary to the RTC and RTCC device
drivers which emulate "dumb" hardware which has no control over,
or knowledge of, the environment they exist in.

The application could for example get the real time from the host
system using native_rtc, and use this time to configure the
emulated RTC and RTCC devices.

Signed-off-by: Bjarki Arge Andreasen <baa@trackunit.com>