samples: drivers: gnss: fix vmu_rt1170 board overlay
The overlay for the vmu_rt1170 incorrectly sets the gnss alias to
the uart node the gnss is on, and is not using the proper node
naming scheme for the u_blox_m10 node. Additionally, the overlay
needs to be updated to match the hwm2 path for the board and soc
vmu_rt1170_mimxrt1176_cm7.

Signed-off-by: Bjarki Arge Andreasen <bjarki@arge-andreasen.me>